moved to path.parse_type_path
