+++ layered_vision/utils/path.py
@@ -1,0 +1,9 @@
+import re
+from typing import Tuple
+
+
+def parse_type_path(path: str) -> Tuple[str, str]:
+    m = re.match(r'^([a-z]+)(:(([^/]|/[^/]).*|))?$', path)
+    if m:
+        return m.group(1), m.group(3) or ''
+    return None, path
